rename shamrock reader + remove unused functions.
For now the header of Shamrock's vtk dumps does not contain the particle mass (which made sense until now, as we were not sure to keep the vtk format anyways). I will do an update of the header format in Shamrock to include all the physical parameters, then add a header reader to this file.